finish compare operations

+++ memory.cpp
@@ -1,6 +1,6 @@
 #include "memory.h"
 
-uint16_t Memory::get_page_offset(uint16_t page) {
+uint16_t Memory::get_page_offset(uint8_t page) {
   uint16_t address = page * 0xFF;
   if (page > 0)
     address++;
@@ -14,11 +14,11 @@ void Memory::set_at(uint16_t address, Byte input) {
   data[address] = input;
 }
 
-Byte Memory::get_at_page(uint16_t page, Byte offset) {
+Byte Memory::get_at_page(uint8_t page, Byte offset) {
   return get_at(get_page_offset(page) + offset);
 }
 
-void Memory::set_at_page(uint16_t page, Byte offset, Byte input) {
+void Memory::set_at_page(uint8_t page, Byte offset, Byte input) {
   set_at(get_page_offset(page) + offset, input);
 }
